requirements.txt
streamlit==1.37.1
fastapi==0.112.0
opentelemetry-instrumentation-fastapi==0.47b0
uvicorn==0.30.5
python-dotenv==1.0.1
google-generativeai==0.7.2
requests==2.31.0
requests-oauthlib==2.0.0
pydantic
